Refactor Assets Namespace and Structure
- Moved assets from `NonPrimitive` to `Assets` namespace.
- Updated `GetProfileResource` to use new asset class paths.
- Simplified asset classes by removing redundant code and reorganizing.
- Introduced `CollectionAsset` trait for managing collections of assets.
- Updated asset classes to utilize new `CollectionAsset` trait.
- Removed obsolete asset classes and redundant methods.
- Adjusted related files and classes to reflect namespace changes and improvements.

Changes include:
- Renamed and moved asset classes.
- Updated class constructors and methods.
- Added new asset classes for labels and followers.
- Removed deprecated classes and methods.
