+++ src/Resources/Assets/ByListAsset.php
@@ -1,25 +1,22 @@
 <?php
 
-namespace Atproto\Resources\Assets\NonPrimitive;
+namespace Atproto\Resources\Assets;
 
-use Atproto\Contracts\HTTP\Resources\AssetContract;
-use Atproto\Contracts\HTTP\Resources\ResourceContract;
-use Atproto\Resources\Assets\BaseAsset;
-use Atproto\Resources\Assets\Primitive\DatetimeAsset;
 use Atproto\Resources\BaseResource;
 use Carbon\Carbon;
 
 /**
  * @method string uri()
  * @method string cid()
- * @method CreatorAsset creator()
+ * @method string name()
+ * @method object purpose()
+ * @method string avatar()
  * @method int listItemCount()
- * @method int joinedWeekCount()
- * @method int joinedAllTimeCount()
  * @method LabelsAsset labels()
+ * @method ViewerAsset viewer()
  * @method Carbon indexedAt()
  */
-class JoinedViaStarterPackAsset implements ResourceContract, AssetContract
+trait ByListAsset
 {
     use BaseResource, BaseAsset;
 
@@ -31,8 +28,8 @@ class JoinedViaStarterPackAsset implements ResourceContract, AssetContract
     public function casts(): array
     {
         return [
-            'creator' => CreatorAsset::class,
             'labels' => LabelsAsset::class,
+            'viewer' => ViewerAsset::class,
             'indexedAt' => DatetimeAsset::class,
         ];
     }
