+++ src/Resources/Assets/FollowerAsset.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Atproto\Resources\Assets\NonPrimitive;
+namespace Atproto\Resources\Assets;
 
 use Atproto\Contracts\HTTP\Resources\AssetContract;
 use Atproto\Contracts\HTTP\Resources\ResourceContract;
-use Atproto\Resources\Assets\Primitive\DatetimeAsset;
 
 /**
  * @method string did()
