+++ src/Resources/Assets/KnownFollowersAsset.php
@@ -1,24 +1,25 @@
 <?php
 
-namespace Atproto\Resources\Assets\NonPrimitive;
+namespace Atproto\Resources\Assets;
 
 use Atproto\Contracts\HTTP\Resources\AssetContract;
-use Atproto\Resources\Assets\BaseAsset;
 use Atproto\Resources\BaseResource;
+use GenericCollection\GenericCollection;
+use GenericCollection\Interfaces\TypeInterface;
 
 class KnownFollowersAsset implements AssetContract
 {
     use BaseResource, BaseAsset;
 
-    public function __construct($content)
+    public function __construct(array $content)
     {
         $this->content = $content;
     }
 
-    public function casts(): array
+    protected function casts(): array
     {
         return [
-            'followers' => FollowersAsset::class
+            'followers' => FollowersAsset::class,
         ];
     }
 }
