--- src/Resources/Assets/Primitive/IntAsset.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Atproto\Resources\Assets\Primitive;
-
-use Atproto\Contracts\HTTP\Resources\AssetContract;
-use Atproto\Resources\Assets\BaseAsset;
-
-class IntAsset implements AssetContract
-{
-    use BaseAsset;
-
-    public function cast(): int
-    {
-        return (int) $this->value;
-    }
-}
