--- src/Resources/Assets/Primitive/ObjectAsset.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Atproto\Resources\Assets\Primitive;
-
-use Atproto\Contracts\HTTP\Resources\AssetContract;
-use Atproto\Resources\Assets\BaseAsset;
-
-class ObjectAsset implements AssetContract
-{
-    use BaseAsset;
-
-    public function cast(): object
-    {
-        return (object) $this->value;
-    }
-}
